requirements.txt
slack-bolt>=1.20.0
Flask>=3.0.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
rapidfuzz>=3.0.0

pytest>=8.0.0
